requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
mysqlclient
python-dotenv